donut/requirements.txt
--find-links https://download.pytorch.org/whl/cu111/torch_stable.html
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
datasets==2.15.0
dill==0.3.7
exceptiongroup==1.2.0
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
google-auth==2.25.2
google-auth-oauthlib==1.0.0
gradio==4.8.0
gradio_client==0.7.1
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
lightning-utilities==0.10.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.4
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.15
nltk==3.8.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.1.78
orjson==3.9.10
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.1
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==1.6.4
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
sacremoses==0.1.1
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
timm==0.5.4
tokenizers==0.10.3
tomlkit==0.12.0
toolz==0.12.0
torch==1.10.1
torchaudio==0.10.1
torchmetrics==1.2.1
torchvision==0.11.2
tqdm==4.66.1
transformers==4.11.3
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
websockets==11.0.3
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
zss==1.2.0
